apps/bot/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
uvloop==0.19.0 ; sys_platform != 'win32' and platform_python_implementation == 'CPython'
httpx==0.27.0
